+++ StabilityMatrix.Avalonia/Models/Inference/LatentType.cs
@@ -1,0 +1,8 @@
+﻿namespace StabilityMatrix.Avalonia.Models.Inference;
+
+public enum LatentType
+{
+    Default,
+    Sd3,
+    Hunyuan
+}