+++ StabilityMatrix.Avalonia/ViewModels/CheckpointBrowser/HuggingFacePageViewModel.cs
@@ -155,12 +155,13 @@ public partial class HuggingFacePageViewModel : TabViewModelBase
                 var url =
                     $"https://huggingface.co/{viewModel.Item.RepositoryPath}/resolve/main/{file}?download=true";
                 var sharedFolderType = viewModel.Item.ModelCategory.ConvertTo<SharedFolderType>();
+                var fileName = Path.GetFileName(file);
                 var downloadPath = new FilePath(
                     Path.Combine(
                         Design.IsDesignMode ? string.Empty : settingsManager.ModelsDirectory,
                         sharedFolderType.ToString(),
                         viewModel.Item.Subfolder ?? string.Empty,
-                        file
+                        fileName
                     )
                 );
                 downloadPath.Directory?.Create();