+++ StabilityMatrix.Avalonia/ViewModels/Inference/InferenceWanImageToVideoViewModel.cs
@@ -2,13 +2,10 @@
 using System.Text.Json.Serialization;
 using Injectio.Attributes;
 using StabilityMatrix.Avalonia.Models;
-using StabilityMatrix.Avalonia.Models.Inference;
 using StabilityMatrix.Avalonia.Services;
 using StabilityMatrix.Avalonia.ViewModels.Base;
-using StabilityMatrix.Avalonia.ViewModels.Inference.Video;
 using StabilityMatrix.Avalonia.Views.Inference;
 using StabilityMatrix.Core.Attributes;
-using StabilityMatrix.Core.Models;
 using StabilityMatrix.Core.Models.Api.Comfy.Nodes;
 using StabilityMatrix.Core.Services;
 
@@ -16,38 +13,8 @@ namespace StabilityMatrix.Avalonia.ViewModels.Inference;
 
 [View(typeof(InferenceWanImageToVideoView), IsPersistent = true)]
 [RegisterTransient<InferenceWanImageToVideoViewModel>, ManagedService]
-public class InferenceWanImageToVideoViewModel : InferenceGenerationViewModelBase, IParametersLoadableState
+public class InferenceWanImageToVideoViewModel : InferenceWanTextToVideoViewModel
 {
-    private readonly ServiceManager<ViewModelBase> vmFactory;
-    private readonly IInferenceClientManager inferenceClientManager;
-    private readonly INotificationService notificationService;
-    private readonly ISettingsManager settingsManager;
-    private readonly RunningPackageService runningPackageService;
-
-    [JsonIgnore]
-    public StackCardViewModel StackCardViewModel { get; }
-
-    [JsonPropertyName("Model")]
-    public WanModelCardViewModel ModelCardViewModel { get; }
-
-    [JsonPropertyName("Sampler")]
-    public SamplerCardViewModel SamplerCardViewModel { get; }
-
-    [JsonPropertyName("BatchSize")]
-    public BatchSizeCardViewModel BatchSizeCardViewModel { get; }
-
-    [JsonPropertyName("Seed")]
-    public SeedCardViewModel SeedCardViewModel { get; }
-
-    [JsonPropertyName("ImageLoader")]
-    public SelectImageCardViewModel SelectImageCardViewModel { get; }
-
-    [JsonPropertyName("Prompt")]
-    public PromptCardViewModel PromptCardViewModel { get; }
-
-    [JsonPropertyName("VideoOutput")]
-    public VideoOutputSettingsCardViewModel VideoOutputSettingsCardViewModel { get; }
-
     public InferenceWanImageToVideoViewModel(
         ServiceManager<ViewModelBase> vmFactory,
         IInferenceClientManager inferenceClientManager,
@@ -57,54 +24,18 @@ public class InferenceWanImageToVideoViewModel : InferenceGenerationViewModelBas
     )
         : base(vmFactory, inferenceClientManager, notificationService, settingsManager, runningPackageService)
     {
-        this.vmFactory = vmFactory;
-        this.inferenceClientManager = inferenceClientManager;
-        this.notificationService = notificationService;
-        this.settingsManager = settingsManager;
-        this.runningPackageService = runningPackageService;
-
-        SeedCardViewModel = vmFactory.Get<SeedCardViewModel>();
-        SeedCardViewModel.GenerateNewSeed();
-
-        ModelCardViewModel = vmFactory.Get<WanModelCardViewModel>(vm =>
-        {
-            vm.IsClipVisionEnabled = true;
-        });
-
-        SamplerCardViewModel = vmFactory.Get<WanSamplerCardViewModel>(samplerCard =>
-        {
-            samplerCard.IsDimensionsEnabled = true;
-            samplerCard.IsCfgScaleEnabled = true;
-            samplerCard.IsSamplerSelectionEnabled = true;
-            samplerCard.IsSchedulerSelectionEnabled = true;
-            samplerCard.IsDenoiseStrengthEnabled = true;
-            samplerCard.DenoiseStrength = 1.0d;
-            samplerCard.EnableAddons = false;
-        });
-
-        PromptCardViewModel = AddDisposable(vmFactory.Get<PromptCardViewModel>());
-
-        BatchSizeCardViewModel = vmFactory.Get<BatchSizeCardViewModel>();
-
         SelectImageCardViewModel = vmFactory.Get<SelectImageCardViewModel>();
-
-        VideoOutputSettingsCardViewModel = vmFactory.Get<VideoOutputSettingsCardViewModel>();
-
-        StackCardViewModel = vmFactory.Get<StackCardViewModel>();
-        StackCardViewModel.AddCards(
-            ModelCardViewModel,
-            SamplerCardViewModel,
-            SeedCardViewModel,
-            BatchSizeCardViewModel,
-            VideoOutputSettingsCardViewModel
-        );
+        SamplerCardViewModel.IsDenoiseStrengthEnabled = true;
+        SamplerCardViewModel.Width = 512;
+        SamplerCardViewModel.Height = 512;
     }
+
+    [JsonPropertyName("ImageLoader")]
+    public SelectImageCardViewModel SelectImageCardViewModel { get; }
 
     /// <inheritdoc />
     protected override void BuildPrompt(BuildPromptEventArgs args)
     {
-        base.BuildPrompt(args);
-
         var builder = args.Builder;
 
         builder.Connections.Seed = args.SeedOverride switch
@@ -148,93 +79,5 @@ public class InferenceWanImageToVideoViewModel : InferenceGenerationViewModelBas
         {
             yield return image;
         }
-    }
-
-    /// <inheritdoc />
-    protected override async Task GenerateImageImpl(
-        GenerateOverrides overrides,
-        CancellationToken cancellationToken
-    )
-    {
-        if (!await CheckClientConnectedWithPrompt() || !ClientManager.IsConnected)
-        {
-            return;
-        }
-
-        if (!await ModelCardViewModel.ValidateModel())
-            return;
-
-        // If enabled, randomize the seed
-        var seedCard = StackCardViewModel.GetCard<SeedCardViewModel>();
-        if (overrides is not { UseCurrentSeed: true } && seedCard.IsRandomizeEnabled)
-        {
-            seedCard.GenerateNewSeed();
-        }
-
-        var batches = BatchSizeCardViewModel.BatchCount;
-
-        var batchArgs = new List<ImageGenerationEventArgs>();
-
-        for (var i = 0; i < batches; i++)
-        {
-            var seed = seedCard.Seed + i;
-
-            var buildPromptArgs = new BuildPromptEventArgs { Overrides = overrides, SeedOverride = seed };
-            BuildPrompt(buildPromptArgs);
-
-            // update seed in project for batches
-            var inferenceProject = InferenceProjectDocument.FromLoadable(this);
-            if (inferenceProject.State?["Seed"]?["Seed"] is not null)
-            {
-                inferenceProject = inferenceProject.WithState(x => x["Seed"]["Seed"] = seed);
-            }
-
-            var generationArgs = new ImageGenerationEventArgs
-            {
-                Client = ClientManager.Client,
-                Nodes = buildPromptArgs.Builder.ToNodeDictionary(),
-                OutputNodeNames = buildPromptArgs.Builder.Connections.OutputNodeNames.ToArray(),
-                Parameters = SaveStateToParameters(new GenerationParameters()) with
-                {
-                    Seed = Convert.ToUInt64(seed)
-                },
-                Project = inferenceProject,
-                FilesToTransfer = buildPromptArgs.FilesToTransfer,
-                BatchIndex = i,
-                // Only clear output images on the first batch
-                ClearOutputImages = i == 0
-            };
-
-            batchArgs.Add(generationArgs);
-        }
-
-        // Run batches
-        foreach (var args in batchArgs)
-        {
-            await RunGeneration(args, cancellationToken);
-        }
-    }
-
-    /// <inheritdoc />
-    public void LoadStateFromParameters(GenerationParameters parameters)
-    {
-        SamplerCardViewModel.LoadStateFromParameters(parameters);
-        ModelCardViewModel.LoadStateFromParameters(parameters);
-        PromptCardViewModel.LoadStateFromParameters(parameters);
-        VideoOutputSettingsCardViewModel.LoadStateFromParameters(parameters);
-        SeedCardViewModel.Seed = Convert.ToInt64(parameters.Seed);
-    }
-
-    /// <inheritdoc />
-    public GenerationParameters SaveStateToParameters(GenerationParameters parameters)
-    {
-        parameters = SamplerCardViewModel.SaveStateToParameters(parameters);
-        parameters = ModelCardViewModel.SaveStateToParameters(parameters);
-        parameters = PromptCardViewModel.SaveStateToParameters(parameters);
-        parameters = VideoOutputSettingsCardViewModel.SaveStateToParameters(parameters);
-
-        parameters.Seed = (ulong)SeedCardViewModel.Seed;
-
-        return parameters;
     }
 }