+++ StabilityMatrix.Avalonia/Views/InferencePage.axaml.cs
@@ -62,6 +62,11 @@ public partial class InferencePage : UserControlBase
         (DataContext as InferenceViewModel)!.AddTabCommand.Execute(InferenceProjectType.ImageToVideo);
     }
 
+    private void AddTabMenu_WanTextToVideo_OnClick(object? sender, RoutedEventArgs e)
+    {
+        (DataContext as InferenceViewModel)!.AddTabCommand.Execute(InferenceProjectType.WanTextToVideo);
+    }
+
     private void AddTabMenu_WanImgToVideo_OnClick(object? sender, RoutedEventArgs e)
     {
         (DataContext as InferenceViewModel)!.AddTabCommand.Execute(InferenceProjectType.WanImageToVideo);